refactor remove duplicates, move to utils, add tests
