+++ internal/storage/elasticsearch.go
@@ -327,19 +327,6 @@ func (es ElasticSearch) GetAttributes(filter *AttributeFilter, tenantID string) 
 	return unique, nil
 }
 
-// RemoveDuplicates removes duplicates from a slice of strings while preserving the order.
-func RemoveDuplicates(s []string) []string {
-	seen := make(map[string]struct{}, len(s))
-	var result []string
-	for _, v := range s {
-		if _, ok := seen[v]; !ok {
-			seen[v] = struct{}{}
-			result = append(result, v)
-		}
-	}
-	return result
-}
-
 // MaxLimit grabs the configured maxlimit for results
 func (es ElasticSearch) MaxLimit() uint {
 	return uint(viper.GetInt("elasticsearch.max_result_window"))
