+++ internal/storage/util_test.go
@@ -1,0 +1,44 @@
+package storage
+
+import (
+	"reflect"
+	"testing"
+)
+
+func TestRemoveDuplicates(t *testing.T) {
+	tests := []struct {
+		name     string
+		input    []string
+		expected []string
+	}{
+		{
+			name:     "No duplicates",
+			input:    []string{"apple", "banana", "cherry"},
+			expected: []string{"apple", "banana", "cherry"},
+		},
+		{
+			name:     "With duplicates",
+			input:    []string{"apple", "banana", "cherry", "apple", "cherry"},
+			expected: []string{"apple", "banana", "cherry"},
+		},
+		{
+			name:     "All duplicates",
+			input:    []string{"apple", "apple", "apple"},
+			expected: []string{"apple"},
+		},
+		{
+			name:     "Single element",
+			input:    []string{"apple"},
+			expected: []string{"apple"},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			output := RemoveDuplicates(tt.input)
+			if !reflect.DeepEqual(output, tt.expected) {
+				t.Errorf("got %v, want %v", output, tt.expected)
+			}
+		})
+	}
+}
